Commit 3A2955A5ED4D370A0395B499782654E0F9B5E5A2
- BUGFIX: unneeded appearance of "Deleting event log archives older than 90 days:"
- Code formatting

+++ Start-RoutineCheckupEventLogCleanUp.ps1
@@ -7,24 +7,29 @@ Function Start-RoutineCheckupEventLogCleanUp
 	{
 		Write-Host -Object "OK." -ForegroundColor Green
 
-		Write-Host -Object "Deleting event log archives older than 90 days: " -NoNewline
-		Get-ChildItem -Path (Join-Path -Path $EventLogBackupPath -ChildPath '*.evtx') -Force | Where-Object `
+		$ExpiredEventLogArchive = Get-ChildItem -Path (Join-Path -Path $EventLogBackupPath -ChildPath '*.evtx') -Force | Where-Object `
 		{
 			-not $_.PSIsContainer -and $_.LastWriteTime -lt ([DateTime]::Now - (New-TimeSpan -Days 90))
-		} | `
-		ForEach-Object `
+		}
+
+		if (-not [Object]::ReferenceEquals($NULL, $ExpiredEventLogArchive))
 		{
-			try
+			Write-Host -Object "Deleting event log archives older than 90 days: " -NoNewline
+
+			$ExpiredEventLogArchive | ForEach-Object `
 			{
-				Write-Host -Object "$([Char] 0x2022) Deleting " -NoNewline
-				Write-Host -Object $_.Name -NoNewline -ForegroundColor Cyan
-				Write-Host -Object "... " -NoNewline
-				Remove-Item -Path $_.FullName -Force -ErrorAction Stop
-				Write-Host -Object "Done." -ForegroundColor Green
-			}
-			catch
-			{
-				Write-Host -Object "Failed ($($_.Exception.Message))." -ForegroundColor Red
+				try
+				{
+					Write-Host -Object "$([Char] 0x2022) Deleting " -NoNewline
+					Write-Host -Object $_.Name -NoNewline -ForegroundColor Cyan
+					Write-Host -Object "... " -NoNewline
+					Remove-Item -Path $_.FullName -Force -ErrorAction Stop
+					Write-Host -Object "Done." -ForegroundColor Green
+				}
+				catch
+				{
+					Write-Host -Object "Failed ($($_.Exception.Message))." -ForegroundColor Red
+				}
 			}
 		}
 	}
@@ -47,7 +52,7 @@ Function Start-RoutineCheckupEventLogCleanUp
 	if (Test-Path -Path $EventLogBackupPath -PathType Container)
 	{
 		$ArchivingTime = [DateTime]::Now
-		
+
 		Write-Host -Object "Enumerating event logs... " -NoNewline
 		$EventLogs = Get-WmiObject -Class Win32_NTEventLogFile
 		Write-Host -Object "Done." -ForegroundColor Green
